Correct associativity for infix:<Z>
This commit corrects the associativity for the `infix:<Z>` operator
from `assoc<chain>` to `assoc<list>`.  `Z` has list associativity
because `$a Z $b Z $c == infix:<Z>($a, $b, $c)`.  It does not have
chain associativity: `($a Z $b Z $c) ≠ ($a Z $b) and ($b Z $c)`.

The chart at the top of the page already correctly indicated that `Z`
has list associativity and thus does not need to be updated.