Add link from containers to ordered/unordered role docs
I was surprised just now when I noticed the Containers doc has no
links to either "Lists, sequences and arrays" or "Sets, bags and
mixes".

As a wayfinding problem, I don't see a better place for it than here
in a new final paragraph of the introduction.

(Checking other VHLL's docs, they seem mostly to cover the
subjects of all three of our docs in a "collections" or "abstract data
types" overview discussion. I don't think Raku's suited to this
treatment and the three docs as they are make sense. But this one
should point at the other two.)