Allow multiline input in `ignored-methods.txt`
The previous version had an honestly very cool bit of code that parsed
`ignored-methods.txt` into a lazily evaluated hash.  This commit
removes that code, sadly, though doing so doesn't have a runtime
penalty (and if `ignored-methods.txt` ever gets large enough that
there's a large runtime cost to parsing it, we have other problems…).

I changed approaches for two reasons: First (and most importantly),
the old approach did not allow for multiline input, which made it very
difficult to document why specific methods should be ignored.  Second,
the old approach used a custom syntax that, while simple, wasn't clear
from context.

The new implementation is for the user to specify methods to ignore in
a standard Raku hash, which both allows for comments and should
provide a syntax that's familiar to all contributors.