Improve output presentation and program performance
This commit avoids printing output in the body of the main program
execution, which allows the output to be presented more logically/in a
better order.  At the same time, it avoids waiting to print anything
until after all output has been produced, which would significantly
increase the time required to begin producing output when scanning
large directories.  To strike the right balance, this revision makes
use of a lazy list of all output and prints results from that list as
they are processed.