Add update-html target
... to update the HTML.  This is easier to remember than the
`documentable update` command as it uses the make infrastructure which
is already present.  It also uses the same general interface to the
various build processes provided by make (i.e. one can use `make
<target>` as opposed to having to remember what command is necessary to
perform a particular operation).  Also, this wraps the `update`
process behind an interface; the details of which can be changed in the
future without changing the interface.