Add `update-html` make target (#3544)
* Add update-html target

... to update the HTML.  This is easier to remember than the
`documentable update` command as it uses the make infrastructure which
is already present.  It also uses the same general interface to the
various build processes provided by make (i.e. one can use `make
<target>` as opposed to having to remember what command is necessary to
perform a particular operation).  Also, this wraps the `update`
process behind an interface; the details of which can be changed in the
future without changing the interface.

* Use update-html make target in README

* Simplify docs for HTML generation

The `html` make target has `for-documentable` as a dependency and then
runs the `documentable` command mentioned in the documentation,
therefore both of these commands can be replaced by `make html` in the
README.

* Remove superfluous vertical whitespace

The problem I see here is that we're using in a local Makefile a tool that's not a dependency. Spinning off all web generation tools is in the roadmap, but we haven't got around to it. Ideally, we should have a repo that has documentable _and_ doc as a dependency, and can be used to build the documentation (including goodies like automatic deployment). Meanwhile, this is perfect. Thanks.