Improve --exclude option & add inverse --only option
This commit updates the --exclude option to take a Regex rather than
a comma-separated list of file names.  No files or directories that
match that Regex will be scanned.  This also adds an --only option,
which also takes a Regex; only files/directories that match the Regex
will be scanned.  The two options can be used together to display only
files that match --only and do not match --exclude.