Replace dynamic variable hack in fmt-with-percent
Previously, this script used a dynamic variable in the
fmt-with-percent sub to allow it to access both the name and value of
a variable when passed a string with the name of that variable.  (The
purpose of this was to allow the sub to print something like "22
percent of total" when passed 'total' in a scope with a $total
variable).  This worked, but was a bit of a hack and introduced
needless complexity.

Now, the sub takes a named argument slurpy constrained to contain only
one key.  This allows the caller to pass in any one named argument,
and lets the sub access both the name (key) and value of that
argument.  Thus we get the same behavior on the calling
side (actually, even better, since we can pass in a named argument
instead of a string) and avoid the need for dynamic variables.