Add --exclude-dir and --only-dir CLI options
Previously, the --exclude and --only options applied to both files and
directories.  This led to counter-intuitive behavior, such as
specifying --only needing to match *both* the filename *and* the
directory name in order to scan a file.

Now, the user can specify filters for files and directories
separately, which should make the script much easier to use.