Refactor summary-formatting code
Previously, the Summary object tracked redundant information: it
tracked both elements in various categories and, separately, the total
number of elements across all categories.  However, the second can be
trivially computed as the sum of the first; tracking them separately
prevents having a single source of truth and created an opportunity
for bugs to hide.  This commit removes that opportunity (and fixes at
least one such bug).

It also cleans up/refactors the summary formatting code in a few
other, more minor ways.