Use sum type for error handling
This commit adds a Result sum type (inspired by Rust's Result).  This
type is either an Ok or a Err type; the calling function must match on
the type and handle any errors before calling `.unwrap` to get the
inner value.  Using the Result type significantly simplifies handling
the error cases when we cannot check the type.

My approach to creating the Result type was built on the discussion at
https://wimvanderbauwhede.github.io/articles/roles-as-adts-in-raku/