Handle missing/invalid --ignore file
Previously, the script could not function if not provided with a
file listing ignored methods (or allowed to use the default).  Now, it
will work (with a warning) if provided a non-existent file (including
the empty string) and will end execution with a descriptive error if
given an existing file that cannot be parsed.