Provide a length hint for aspell
Giving aspell a hint about the length of the word-list we provide it
allows it to perform very slightly (but measurably) faster.  This
appears to improve the speed of our spelling test by ~2%.  The hint
does not need to be perfectly accurate to be helpful, so it won't be a
problem if the hint isn't updated as contributors add more words to
the aspell word list.

See http://aspell.net/man-html/Format-of-the-Personal-and-Replacement-Dictionaries.html#Format-of-the-Personal-and-Replacement-Dictionaries