First attempt at Distribution::Resource
This refs mainly #3629, although it also tries to adress #3043.
It is a functional description, rather than a syntactic description; as a matter of fact, it's probably not sensible to describe one by one all the "IO" look-alike methods it's got, but a high-level one.

Not even the introduction is ready, it needs some work:

* Adds usage of "libraries" in "is native"
* Using resources directly.