Don't guess about rakudo checkout.
Move some docs from internal to user-facing>

Force the user to select a directory, warn the user in docs that we will be running git commands in their checkout.

s/repository/checkout/ for clarity in the error message, show error if var isn't set or if directory isn't there.

Fix directory check - previously would allow a file to be passed as RAKUDO_SRC

Closes #3679