Update Programs/03-environment-variables.pod6
This file is a copy of [rakudo/docs/running.pod]<https://github.com/rakudo/rakudo/blob/master/docs/running.pod>
but ran out of sync. Merge the two files. Also now include all bits of the
rakudo/rakudo version of the file and not only the environment variable
paragraph.

Of note this also pulls in the documentation for the new `rakuw.exe` and
`rakudow.exe` executables.