Remove explicit invocant typing in Metamodel roles
In rakudo, these roles are defined in NQP,
where this typing isn't part of the method definition.
We can add back if determined that this is part of the spec,
and not just an implementaiton detail.

Additionally, these tests now fail to compile under rakudo
which is not by itself reason enough to remove them.