Move the pod precompilation cache
When tracking why a test of a single file was taking 20 seconds to load
Test-Files, nine++ pointed out that a large number of files to scan in
load can cause this kind of slowdown.

The precompilation cache is now in .../.pod-precomp - you can move
your precomp folder there, or just delete it from lib and start over.