Largely rework the Object Tutorial
With the advent of `is built` all the common use-cases of object
construction are now covered by our attribute syntax. This means that the
need to fall back to using `BUILD` has lessened enormously. Given that
`BUILD` has a number of behaviours that are non-obvious and require
knowing and understanding them, it makes sense to consider `BUILD` the
last resort mechanism to use when all else fails and recommend using
automatic attribute initialization and `TWEAK` for everything else.
Communicating this clearly in our object tutorial is very valuable, it
shows newcomers the sane default first. We do not want to send them off in
the wrong direction as `BUILD` has proven to be a stumbling block for new
users numerous times.

(Probably not entirely complete) List of Changes:

- Try to be more structured by reordering things.
- Add more headings for easier navigation.
- Explain `is built`.
- Rework examples to not use `BUILD` anymore.
- Remove wrong explanation of the `Task` example stating one would need a
  `BUILD` when custom constructors are used.
- Remove in-depth explanation of `BUILD`. Add respective documentation for
  `TWEAK`. `BUILD` is still explained in detail in the Object Reference.
- State that `BUILD` is only intended for advanced use-cases.
- Remove `Hero` example which is made obsolete by `is built`.
- Simplify the `Str-with-ID` example and remove bits promoting making
  `TWEAK` a `method` instead of a `submethod`.
- Remove confusing bit about `my` vs `our` class variables.
- Rename `Static Fields?` heading to `Class variables`
- Replace the `Singleton` example with the `Str-with-ID` example as that
  shows the "variable shared by all instances" aspect better than a
  singleton (of which there usually only is one).