Programs/03-Environment... remove F<
`F<...>` is not a standard FormatCode, either in the current Pod6 documentation, or in Damian Conway's original spec. This is the only Raku documentation content file to contain `F<...>`. The current standard HTML renderer just stringifies `F< ... >` to `F< ...>`, which is quite ugly imho. I have replaced `F<...>` with `B<...>`, which is a standard FormatCode.