Remove link from List #3841
This is another false link that probably needs removing.
There is no `routine_Str`or `method_Str` documented in `Any`. I could not find a single reference to `.Str` that I could replace the reference with.
Removing the link does not seem to me to be important, but it probably should be an issue that `Str` is not documented in a single place. Strings and stringification are important enough to need a separate document.