Insert missing target into Proc #3841
`Type/Independent-routines` contains a reference to `Proc#sub_run` but `Type/Proc` does not contain any target for `run`.

So I have inserted a X<> around the first explanation of `run` in Proc. This provides a target for Independent-routines (also opening separate PR with new target)

BUT, for consistency, `run` should be listed as sub along with `new` and `shell`. Also relationship with new should be illustrated. This is not done in this PR.