Improve documentation of the list routine
The documentation of the list subroutine is a good place to explain the
single argument rule yet again.

Most core types implementing a .list method only accept an object instance
as the invocant. If the core type implements .list as a multi method,
listing the type object is handled by

    multi method list(Any:U: --> List)