Remove incorrect text & non-compiling example
This commit removes some text that describes binding semantics that,
after discussion, turned out to be the result of a bug rather than
intended behavior.  It also removes the accompanying test, which
failed to compile once the bug was fixed.

Finally, it adds a doc-comment clarifying that an adjacent test fails
at runtime even though it *looks* like it would throw a compile-time
error (and thus would need to be marked for xt/example-compilation to
skip).