Make grep selection dumber.
Grep and aspell disagree about what a word is; 'compat', for example, isn't found by grep -wli,
but is considered a word because it's separated by underscores.

It means we'll have to test more files to be sure we still need the entry, but removes a chance
for false positives.