Add scrollbar for navbar
The navbar was made sticky in b2e1f85 to avoid situations where the
user needs to scroll the main content up to use the ToC navbar pane.
However, this resulted in some portions of the ToC being unreachable
when the ToC was long enough that it didn't fit on a single screen.

This commit fixes that by adding a scrollbar to the navbar.