Modify the 'Brackets' table generator to read the data source file
The original table generator still has the embedded
data, but its current default is to read the source
file and extract the data so it can be updated if
the source should change.

For now the script allows for other options that
can be removed when the default mode satisfies
the current requirements.

Possible changes:

+ modify the behavior for a failed download of the source
  HLL::Grammar.nqp
+ change the table name
+ change the pod6 file name