Update Parameter.rakudoc
In the edited example, I see no way how the ::U type capture can be different from the ::T one. So it appears that the possibility of having more than one type capture per argument is useless. If an explanation can be provided, I will gladly edit it back in, with the explanation.