Language/modules.rakudoc contains a fatal flaw
The changes introduced on March 29 broke the website :(
The changes to this file are in fact correct Rakudoc. However Rakudo does not compile it correctly.
This can be verified by creating a short file with the changes given here, "bracketing" them with `=begin pod`, `=end pod` directives, and then the raku line `say $=pod.raku`.
The block structure is incorrect, and the `=item` at line 712 is not shown as having level = 2.
This is a bug in Raku. It relates to multilevel lists that contain `=begin item`. 
Multilevel lists like
```
=item top level
=item2 next level
=item3 third level
```
work as expected.

The work around until ProcessedPod is improved is to move the inner item beginning at 712 to a level 1 item. 
This is what this PR does.

Since Raku does not compile it correctly, the ProcessedPod Renderer fails. I am working on ProcessedPod, so that a mal-formed list, like this one, is transformed into some form of HTML without failing completely.

However, for the time being,