Temporarily fix rendering of X<>
Pod::To::Text renders C<X<>> as "X<>", which is incorrect
RakuDoc::To::Text renders C<X<>> as "", which is correct

However the desired result is X<>, and that should be specified as
C<V<X<>>> .  But that renders as V<X<>> in Pod::To::Text, so there
is no way to get monospaced X<> from both Pod:: and RakuDoc::.

Change the markup to V<X<>>: this is rendered correctly by both,
but loses the monospacedness.  Which is livable, I think.