Fix rendering
The old pod grammar doesn't handle 'C«<O(I<...>)>»' well, in the
sense that it doesn't consider the inner 'I<...>' to be markup.

The RakuAST grammar does, causing difference in rendering.  Since
the I<...> is really only for a little emphasis, it feels ok to
remove that for now to get consistent rendering.