Revert / Update to make RakuAST rendering tests pass
This reverts a number of recent commits that changed C<> to use
V<> as it was my impression that only V<> had verbatim semantics.
Turns out this was not the idea:
  https://github.com/finanalyst/Rakudoc-BETA/issues/25#issuecomment-1621448351

Fixed the parser to reflect this decision, reverted all of the C<>
-> V<> changes, and tweaked some of the V<> and C<> codes that had
additional whitespacing: the legacy podification logic would flatten
spaces inside verbatim texts.  The RakuAST podification only changes
newlines to spaces.  This caused fails in the rendering tests.  These
were fixed for now by making sure the contents of verbatim pod in
here only contains single spaces.

All clean again now and closer to the original :-)