Replace incorrect use of "currying" with "priming"
The documentation currently uses the term "currying" to refer to what
is more broadly known as "partial function application" (but
frequently incorrectly called "currying").  However, Raku generally
doesn't use the phrase "partial function application", probably
because it's a real mouthful and could be off-putting to people not
familiar with functional programming jargon.  Instead, Raku prefers
the term "priming", see
[S06](https://design.raku.org/S06.html#Priming).

Accordingly, this commit changes (nearly) all uses of "currying" or
related words to the "priming" equivalent; it also adds a "priming"
entry to the glossary.

The only two (intentional) uses of "currying" left are in a TODO
section of the Haskell-to-Raku guide and in the documentation for
the (Rakudo-specific, unspecified) type `CurriedRoleHOW`.