One way of addressing a formatto in create-cli.rakudoc
There's a segment in the doc which has two consecutive `=item` lists. It's clear to me the second one is supposed to correspond to the first one, with three items in the second list to match the first list. But the second just had two items. This PR just fixes that basic mistake, and is a minimalist approach to addressing that basic problem.

However, t seems to me the two lists would be better as a single list of three items, or perhaps better, a table. Rather than write that (more complicated) PR I've decided to wait for responses to this simpler solution.