Finish s/`<|w>`/`<?wb>`/ et al in regexes.rakudoc
See https://github.com/Raku/doc/pull/4435.

I'm not 100% sure this commit will work smoothly as is. One change (or was it two?) involved what looks like meta information related to search indexes. I'm not familiar with that stuff. I did a mechanical translation and guessed it was likely it would work if there was an index entry for `<?wb>`. But maybe I misunderstand. And anyway, I don't know how to test it short of getting into setting up a whole doc testing system, and that's way beyond my $docgrade.