Finish s/`<|w>`/`<?wb>`/ in doc (headings.rakutest)
While `<|w>` works in Rakudo, it isn't official Raku, and while I personally like it, there's a chance it will go away. Its semantics are the same as `<?wb>`, which _is_ official Raku, so I've switched to that. (See also https://github.com/Raku/doc/commit/5c1a147e68e800a35baea3505f4cdc14fee63efa.)

I meant my prior related commit (which changed the regex doc page) to be a PR because I wasn't 100% sure about the `X<...>` stuff and regardless of that I felt it could still do with a quick check by someone else. But I clicked the wrong button. Otoh, once I'd done that I thought it would almost certainly work and be wanted, and if it didn't it wouldn't be a big deal and would be easy for me to spot it even if no one else does, and to fix it.

But for _this_ change I'm making sure I do it as a PR. I still expect it to work, but it may not, and the consequence of that might be a waste of others' time, and even if it does it might not be wanted.