Update section `trait is rw` in Routine.rakudoc
With Rakudo Star v2025.08, the output of the code example is different than currently stated in the doc page. The apparent reason is that `say %hash.raku` now produces output in a different format. In this specific example, it's

`{:some(${:key($[Any, [Any, Any, "autovivified"]])})}`

Since this is rather unwieldy, I propose changing `say %hash.raku` to `say %hash`, with which the output is

{some => {key => [(Any) [(Any) (Any) autovivified]]}}

The same output would also be produces by `say %hash.gist`.