Further simplify `say` line of `is rw` example in Routine.rakudoc
Imo the change just made was a solid step in the right direction, but didn't go far enough.

I have saved a full rationale in a gist if whoever reads this (@coke?) doesn't agree that the output is still far too complex for good reference doc.

(As a tangent, the routine code in the example is also far too complex for good reference doc for `is rw`. But without reading some large swathe of the doc I wouldn't know whether there's a significant negative impact if we radically simplify that code. So I've focused for this PR on just improving further on the good direction @schultzdavid took the `say` part.)