Update signatures.rakudoc: Change `=:=` to `===` and adapt text
The change of `=:=` to `===` is necessary in the second example, because `=:=` would just always produce False there.

From the explanation above the first example, `===` also seems like the better choice there (even though `=:=` also works as intended). 

Namely, it is `===` which calls the `.WHICH` method. I've added this as a remark to the text, and have rearranged the text slightly.