In signatures.rakudoc, remove mention of nonexistent trait `is optional`, and change `=:=` to `===` (#4645)
* In signatures.rakudoc, remove mention of nonexistent trait `is optional`

Trying to use the trait "is optional" in a function signature throws error "Can't use unknown trait" (in all of v6.c, v6.d and v6.e.PREVIEW).

* Update signatures.rakudoc: Change `=:=` to `===` and adapt text

The change of `=:=` to `===` is necessary in the second example, because `=:=` would just always produce False there.

From the explanation above the first example, `===` also seems like the better choice there (even though `=:=` also works as intended). 

Namely, it is `===` which calls the `.WHICH` method. I've added this as a remark to the text, and have rearranged the text slightly.

* Update signatures.rakudoc: whitespace