In signatures.rakudoc, rearrange and modify the sections Sub-signatures, Captures, and Long names (#4646)
* In signatures.rakudoc, modify the sections Sub-signatures, Captures and Long names

Move the entire section on Long names down, because Sub-signatures and Captures are very closely related to one another.

In section Sub-signatures, change from using `|` to `\`, since captures with `|` are already the topic of the `Captures` section thereafter, and because `\` gives behavior that's more like that of an actual "compound parameter", e.g. that other parameters can still come behind them.

Expand sections Sub-signatures and Captions by examples showing that most of the signature syntax is available for sub-signatures, too.

In section Captures, move the paragraph on unbound captures to the bottom to get a good order of topics.

* semicolons