update signatures of `method contains` and explain handling of Cool params
If $needle and/or $pos is of type Cool, this is no longer directly handled by
a multi of Str.contains, but by a multi of Cool.contains.
Hence, remove such signatures from here and instead add an explanation
and a link to Cool.contains .
Also, order the signatures in a way that's better to digest.